Carga/requirements.txt
pandas==1.3.5
schedule-1.1.0